Feature: Input System (Dushyant Shukla) Added support for polling: 1. Keyboard input 2. Mouse input 3. Cursor position

+++ engine/source/engine/framerate-controller/FramerateController.cpp
@@ -53,7 +53,7 @@ namespace gswy {
 		m_frameTime /= 1000; // converting milliseconds to seconds
 	}
 
-	const float& FramerateController::GetFrameTime() const {
+	const double& FramerateController::GetFrameTime() const {
 		return m_frameTime;
 	}
 }
